add unit tests for controllers and for product whole logic

+++ java-server/src/main/java/org/example/mapper/product/ProductRowMapper.java
@@ -12,7 +12,7 @@ public class ProductRowMapper implements RowMapper<Product> {
     @Override
     public Product mapRow(ResultSet rs, int rowNum) throws SQLException {
         Product product = new Product();
-        product.setId_product(rs.getLong("id_product"));
+        product.setId_product(rs.getInt("id_product"));
         product.setProduct_name(rs.getString("product_name"));
         product.setProduct_characteristics(rs.getString("product_characteristics"));
         product.setCategory_number(rs.getInt("category_number"));
