+++ java-server/src/main/java/org/example/model/product/Product.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Product {
-    private Long id_product;
+    private int id_product;
     private String product_name;
     private String product_characteristics;
     private int category_number;
