+++ java-server/src/main/java/org/example/service/product/ProductService.java
@@ -1,9 +1,8 @@
 package org.example.service.product;
 
+import java.util.List;
 import org.example.dto.product.ProductDto;
 import org.example.dto.product.ProductRequestDto;
-
-import java.util.List;
 
 public interface ProductService {
 
@@ -11,11 +10,11 @@ public interface ProductService {
 
     ProductDto save(ProductRequestDto requestDto);
 
-    ProductDto updateProductById(Long id, ProductRequestDto requestDto);
+    ProductDto updateProductById(int id, ProductRequestDto requestDto);
 
     List<ProductDto> findByName(String name);
 
     List<ProductDto> findByCategoryId(int category_number);
 
-    void deleteProductById(Long id);
+    void deleteProductById(int id);
 }
