+++ java-server/src/main/java/org/example/service/product/ProductServiceImpl.java
@@ -9,8 +9,6 @@ import org.example.repository.product.ProductRepository;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
@@ -35,7 +33,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public ProductDto updateProductById(Long id, ProductRequestDto requestDto) {
+    public ProductDto updateProductById(int id, ProductRequestDto requestDto) {
         return repository.updateProductById(id, requestDto);
     }
 
@@ -56,7 +54,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void deleteProductById(Long id) {
+    public void deleteProductById(int id) {
         repository.deleteById(id);
     }
 }
